[path_provider] Migrate examples to null-safety (#3559)
Allows running the examples in strong mode, even though the integration tests can't yet be.

Converts the macOS example to use the platform interface, rather than the app-facing package, to eliminate the circular dependency. Also does some cleanup and simplification of the desktop example pubspecs.

Does not update versions/changelogs since this won't be explicitly published, given that it's example-only.